+++ src/config.ts
@@ -48,7 +48,7 @@ function cacheForDirs<V>(
 
 let pathToApiMap = expiringMap<string | null, Promise<UnifiedApi>>(10_000)
 
-export async function getTailwindConfig(options: ParserOptions): Promise<any> {
+export async function getTailwindConfig(options: ParserOptions): Promise<UnifiedApi> {
   let cwd = process.cwd()
 
   // Locate the file being processed